Move IntoIter to a module
